requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altgraph==0.17.4
attrs==24.2.0
auto-py-to-exe==2.44.2
backoff==2.2.1
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
Eel==0.17.0
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
multidict==6.1.0
opencage==3.0.3
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pygame==2.6.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.4.0
stepic==0.5.0
tkinterweb==3.24.7
tqdm==4.67.0
urllib3==2.2.3
whichcraft==0.6.1
xml-python==0.4.3
yarl==1.18.0
zope.event==5.0
zope.interface==7.1.1
